style: error page / settings navigation
