+++ src/app/settings/account/page.tsx
@@ -4,16 +4,16 @@ import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
 import { redirect, useRouter } from "next/navigation";
 import AccountHeader from "@/components/account/account-header/account-header";
+import Link from "next/link";
 
 export default function Page() {
   const { data: session } = useSession();
-  const router = useRouter();
 
   if (!session) redirect("/");
 
   return (
     <>
-      <AccountHeader back="/" label="Account" />
+      <AccountHeader back="/settings" label="Account" />
       <div className="flex flex-col gap-3 justify-center items-center">
         <Avatar className="size-20">
           <AvatarImage src={session.user?.image as string} />
@@ -24,21 +24,17 @@ export default function Page() {
 
         <div className="font-mono">{session.user?.email}</div>
 
-        <Button
-          className="w-full"
-          variant={"secondary"}
-          onClick={() => router.push("/account/save")}
-        >
-          Save (merge)
-        </Button>
+        <Link href={"./account/save"} className="w-full">
+          <Button className="w-full" variant={"secondary"}>
+            Save
+          </Button>
+        </Link>
 
-        <Button
-          className="w-full"
-          variant={"secondary"}
-          onClick={() => router.push("/account/load")}
-        >
-          Load (overwrite)
-        </Button>
+        <Link href={"./account/load"} className="w-full">
+          <Button className="w-full" variant={"secondary"}>
+            Load
+          </Button>
+        </Link>
 
         <Button
           className="w-full"
