+++ src/app/settings/account/save/page.tsx
@@ -2,15 +2,14 @@
 import AccountHeader from "@/components/account/account-header/account-header";
 import { Button } from "@/components/ui/button";
 import { Card } from "@/components/ui/card";
-import { useRouter } from "next/navigation";
+import Link from "next/link";
 
 export default function Page() {
-  const router = useRouter();
   return (
     <>
-      <AccountHeader back="/account" label="Save data" />
+      <AccountHeader back="./" label="Save data" />
 
-      <Card className="p-3 bg-secondary">
+      <Card className="p-3 bg-secondary/10">
         <p>
           Do you want <span className="text-green-700">save</span> your account
           data on the <span className="text-blue-700">cloud</span>?
@@ -20,7 +19,9 @@ export default function Page() {
         </p>
 
         <div className=" flex gap-2 w-full justify-center mt-5">
-          <Button onClick={() => router.push("/account")}>Back</Button>
+          <Link href={"./"}>
+            <Button>Back</Button>
+          </Link>
           <Button>Continue</Button>
         </div>
       </Card>
